# docs/swagger.yaml
definitions:
...
  pod.CreatePodContainerRequest:
    properties:
      command:
        items:
          type: string
        type: array
      env:
        items:
          type: string
        type: array
      image:
        type: string
      mount:
        items:
          type: string
        type: array
      name:
        type: string
      network:
        type: string
      port:
        items:
          type: string
        type: array
      tty:
        type: boolean
    type: object
  pod.CreatePodRequest:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      containers:
        items:
          $ref: '#/definitions/pod.CreatePodContainerRequest'
        type: array
      ipcNS:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      networkNS:
        type: string
      uid:
        type: string
      userNS:
        type: string
      utsNS:
        type: string
    type: object
  pod.ScaleReplicaSetRequest:
    properties:
      replicas:
        type: integer
    type: object
...
  utils.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/pods:
    get:
      description: list pod sandbox
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: list pods
      tags:
      - pods
    post:
      consumes:
      - application/json
      description: create a pod sandbox (no container start)
      parameters:
      - description: Pod Spec
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pod.CreatePodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: create pod sandbox
      tags:
      - pods
  /v1/pods/{podId}:
    delete:
      description: remove a pod sandbox
      parameters:
      - description: Pod ID
        in: path
        name: podId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: remove pod sandbox
      tags:
      - pods
    get:
      description: get pod sandbox detail
      parameters:
      - description: Pod ID
        in: path
        name: podId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get pod detail
      tags:
      - pods
  /v1/pods/{podId}/actions/start:
    post:
      description: start a pod sandbox
      parameters:
      - description: Pod ID
        in: path
        name: podId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: start pod sandbox
      tags:
      - pods
  /v1/pods/{podId}/actions/stop:
    post:
      description: stop a pod sandbox
      parameters:
      - description: Pod ID
        in: path
        name: podId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: stop pod sandbox
      tags:
      - pods
...
  /v1/replicasets/{replicaSetId}:
    delete:
      description: remove replica set
      parameters:
      - description: ReplicaSet ID
        in: path
        name: replicaSetId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: remove replica set
      tags:
      - replicasets
    get:
      description: get replica set detail
      parameters:
      - description: ReplicaSet ID
        in: path
        name: replicaSetId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get replica set detail
      tags:
      - replicasets
  /v1/replicasets/{replicaSetId}/actions/scale:
    post:
      consumes:
      - application/json
      description: scale replica set replicas
      parameters:
      - description: ReplicaSet ID
        in: path
        name: replicaSetId
        required: true
        type: string
      - description: Scale Options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pod.ScaleReplicaSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: scale replica set
      tags:
      - replicasets
swagger: "2.0"